fix: duplicate routes, hardcoded configs and duplicate lifespan
