+++ .env.example
@@ -18,8 +18,12 @@ BLOB_STORAGE_TYPE=""                                                            
 BLOB_CONNECTION_STR="DefaultEndpointsProtocol=https;AccountName=?;AccountKey=?;EndpointSuffix=?"    # azure example
 BLOB_STORAGE_CONTAINER_NAME=""
 
-# Message Broker info
+# Message Broker
 MESSAGE_BROKER_URL="amqp://guest:guest@localhost:5672/"
+OCR_QUEUE_NAME="ocr_queue"
+
+# Upload constraints
+MAX_UPLOAD_SIZE_MB=50
 
 # JWT Config
 ACCESS_TOKEN_EXPIRE_MINUTES=1
