+++ app/core/config.py
@@ -1,5 +1,3 @@
-from typing import List
-
 from pydantic_settings import BaseSettings, SettingsConfigDict
 from pydantic import computed_field
 from urllib.parse import quote_plus
@@ -7,7 +5,13 @@ from urllib.parse import quote_plus
 
 class Settings(BaseSettings):
     ENVIRONMENT: str = "production"
-    CORS_ORIGINS: List[str]
+
+    MONGO_USER: str
+    MONGO_PASS: str
+    MONGO_HOST: str
+    MONGO_PORT: int = 10260
+    MONGO_DB_NAME: str = "sdmsdb"
+    MONGO_TLS_INSECURE: bool = True
 
     SQL_SERVER: str
     SQL_DB_NAME: str
@@ -18,20 +22,6 @@ class Settings(BaseSettings):
     SQL_CONNECT_TIMEOUT: int = 60
     SQL_MAX_RETRIES: int = 3
     SQL_RETRY_DELAY_BASE: int = 2
-
-    BLOB_STORAGE_TYPE: str
-    BLOB_CONNECTION_STR: str
-    BLOB_STORAGE_CONTAINER_NAME: str
-    MAX_SIZE_UPLOAD: int = 50 * 1024 * 1024
-
-    MESSAGE_BROKER_URL: str
-
-    MONGO_USER: str
-    MONGO_PASS: str
-    MONGO_HOST: str
-    MONGO_PORT: int = 10260
-    MONGO_DB_NAME: str = "sdmsdb"
-    MONGO_TLS_INSECURE: bool = True
 
     # JWT configs
     ACCESS_TOKEN_EXPIRE_MINUTES: int
@@ -60,6 +50,27 @@ class Settings(BaseSettings):
 
     # Azure Blob Storage (original file storage)
     AZURE_BLOB_CONTAINER_URL: str = ""
+    BLOB_STORAGE_TYPE: str = "azure"
+    BLOB_CONNECTION_STR: str = ""
+    BLOB_STORAGE_CONTAINER_NAME: str = ""
+
+    # Message Broker
+    MESSAGE_BROKER_URL: str = ""
+    OCR_QUEUE_NAME: str = "ocr_queue"
+
+    # CORS
+    CORS_ORIGINS: list[str] = [
+        "http://localhost:5173",
+        "http://127.0.0.1:5173",
+    ]
+
+    # Upload constraints
+    MAX_UPLOAD_SIZE_MB: int = 50
+    SUPPORTED_EXTENSIONS: set[str] = {
+        ".pdf", ".jpg", ".jpeg", ".png",
+        ".tif", ".tiff", ".bmp", ".heif", ".heic",
+        ".docx", ".xlsx", ".pptx",
+    }
 
     # FAISS Vector Store
     FAISS_INDEX_DIR: str = "data/faiss"
