+++ pyproject.toml
@@ -5,12 +5,8 @@ description = "NassaQ Python infrastructure/backend"
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
-    "aio-pika>=9.5.8",
-    "aiohttp>=3.13.3",
     "aioodbc>=0.5.0",
     "faiss-cpu>=1.9.0",
-    "azure-storage-blob>=12.28.0",
-    "azure-servicebus>=7.14.2",
     "fastapi>=0.128.0",
     "openai>=1.30.0",
     "passlib[bcrypt]>=1.7.4",
